% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_comparison.R
\name{comparison_ops}
\alias{comparison_ops}
\alias{>=<-}
\alias{><-}
\alias{<=<-}
\alias{<<-}
\alias{==<-}
\alias{!=<-}
\alias{\%subset>=\%}
\alias{\%subset>\%}
\alias{\%subset<=\%}
\alias{\%subset<\%}
\alias{\%subset==\%}
\alias{\%subset!=\%}
\title{Comparison Operators}
\usage{
>=(e1, e2) <- value

>(e1, e2) <- value

<=(e1, e2) <- value

<(e1, e2) <- value

==(e1, e2) <- value

!=(e1, e2) <- value

e1 \%subset>=\% e2

e1 \%subset>\% e2

e1 \%subset<=\% e2

e1 \%subset<\% e2

e1 \%subset==\% e2

e1 \%subset!=\% e2
}
\arguments{
\item{e1}{lhs}

\item{e2}{rhs}

\item{value}{replacement value}
}
\description{
Comparison Operators
}
