% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_subset.R
\name{comparison_subset}
\alias{comparison_subset}
\alias{\%[>=\%}
\alias{\%[>\%}
\alias{\%[<=\%}
\alias{\%[<\%}
\alias{\%[==\%}
\alias{\%[!=\%}
\title{Subsetting Values by Comparison}
\usage{
x \%[>=\% y

x \%[>\% y

x \%[<=\% y

x \%[<\% y

x \%[==\% y

x \%[!=\% y
}
\arguments{
\item{x}{first element of the operation.}

\item{y}{second element of the operation.}
}
\value{
elements of \code{x} matched by the used comparison.
}
\description{
Operators for subsetting values using the standard comparison operators.
}
\examples{
ages <- c(130, 10, 21, 39, 74, -2, 0)
ages \%[<\% 5

letters \%[==\% "a"

letters \%[!=\% "a"

}
\seealso{
\code{`==`}
}
