% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_in_elementwise.R
\name{element_wise_in}
\alias{element_wise_in}
\alias{\%in{}\%}
\alias{\%in{}\%<-}
\alias{\%subset{}\%}
\alias{\%!in{}\%}
\alias{\%!in{}\%<-}
\alias{\%!subset{}\%}
\title{Element wise in}
\usage{
x \%in{}\% table

\%in{}\%(x, table) <- value

x \%subset{}\% table

x \%!in{}\% table

\%!in{}\%(x, table) <- value

x \%!subset{}\% table
}
\arguments{
\item{x}{vector or \code{NULL}: the values to be matched.
    \link{Long vectors} are supported.}

\item{table}{vector or \code{NULL}: the values to be matched against.
    \link{Long vectors} are not supported.}

\item{value}{replacement value}
}
\description{
Element wise in
}
