% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_detect.R
\name{\%out\%}
\alias{\%out\%}
\title{Detect values that don't match}
\usage{
x \%out\% table
}
\description{
\code{\%out\%} is the negation of \code{\%in\%}, so typing \code{x \%out\% y} is equivalent to
\code{! x \%in\% y}
}
