% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_replace.R
\name{in_replace}
\alias{in_replace}
\alias{\%in{}\%<-}
\alias{\%!in{}\%<-}
\alias{\%in[]\%<-}
\alias{\%!in[]\%<-}
\alias{\%in()\%<-}
\alias{\%!in()\%<-}
\alias{\%in(]\%<-}
\alias{\%!in(]\%<-}
\alias{\%in[)\%<-}
\alias{\%!in[)\%<-}
\title{Replacing Values and Intervals}
\usage{
x %in{}% table <- value

x %!in{}% table <- value

x %in[]% table <- value

x %!in[]% table <- value

x %in()% table <- value

x %!in()% table <- value

x %in(]% table <- value

x %!in(]% table <- value

x %in[)% table <- value

x %!in[)% table <- value
}
\arguments{
\item{x}{vector or array of values to be matched.}

\item{table}{vector or list to be matched against.}

\item{value}{replacement value.}

\item{interval}{numeric vector defining a range to be matched against.}
}
\value{
\code{x} with specified values replaced with \code{value}.
}
\description{
Operators for replacing values within a given interval or set.
}
\details{
Parentheses of each operator define the type of matching template:\cr
\code{\%in\{\}\%} - set of all elements that will be matched against\cr
\code{\%in[]\%} - closed numeric interval\cr
\code{\%in()\%} - open numeric interval\cr
\code{\%in(]\%} - interval that is is open on the left and closed on the right\cr
\code{\%in[)\%} - interval that is closed on the left and open on the right\cr
Operators starting with a bang (i.e. \code{\%!in\{\}\%}) specify negation.

Operations using numeric intervals are a convenient short hand form of\cr\code{x[x > interval[1] & x < interval[2]] <- value}.

\code{interval} can be specified as a numeric vector of any length.
In such cases \code{range} will be used in order to obtain the endpoints.
}
\seealso{
\code{\%in{}\%}
}
