% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_subset.R
\name{in_subset}
\alias{in_subset}
\alias{\%[in{}\%}
\alias{\%[out{}\%}
\alias{\%[in[]\%}
\alias{\%[out[]\%}
\alias{\%[in()\%}
\alias{\%[out()\%}
\alias{\%[in(]\%}
\alias{\%[out(]\%}
\alias{\%[in[)\%}
\alias{\%[out[)\%}
\alias{\%[in~\%}
\alias{\%[out~\%}
\alias{\%[in~p\%}
\alias{\%[out~p\%}
\alias{\%[in~f\%}
\alias{\%[out~f\%}
\alias{\%[in\%}
\alias{\%[out\%}
\title{Subsetting Values and Intervals}
\usage{
x \%[in{}\% table

x \%[out{}\% table

x \%[in[]\% interval

x \%[out[]\% interval

x \%[in()\% interval

x \%[out()\% interval

x \%[in(]\% interval

x \%[out(]\% interval

x \%in[)\% interval <- value

x \%[out[)\% interval

x \%[in~\% pattern

x \%[out~\% pattern

x \%[in~p\% pattern

x \%[out~p\% pattern

x \%[in~f\% pattern

x \%[out~f\% pattern

x \%[in\% table

x \%[out\% table
}
\arguments{
\item{x}{vector or array of values to be matched.}

\item{table}{vector or list to be matched against.}

\item{interval}{numeric vector defining a range to be matched against.
\code{as.numeric()} is used on lists, it will succeed on flat lists but fail on
nested lists}

\item{pattern}{pattern to be matched against.}
}
\value{
\code{x} with specified values replaced with \code{value}.
}
\description{
Operators for replacing values within a given interval or set.
}
\details{
For each \code{\%[*\%} operator of this package \code{x \%[*\% y} is a shorthand for
\code{x[x \%*\% y]}.
}
\seealso{
\code{\%in{}\%}
}
