% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_in.R
\name{in_variants}
\alias{in_variants}
\alias{\%in\%<-}
\alias{\%!in\%}
\alias{\%!in\%<-}
\title{in variants}
\usage{
\%in\%(x, table) <- value

x \%!in\% table

\%!in\%(x, table) <- value
}
\arguments{
\item{x}{vector or \code{NULL}: the values to be matched.
    \link{Long vectors} are supported.}

\item{table}{vector or \code{NULL}: the values to be matched against.
    \link{Long vectors} are not supported.}

\item{value}{replacement value}
}
\description{
in variants
}
