% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/like_check.R
\name{like_check}
\alias{like_check}
\alias{\%like\%}
\alias{\%!like\%}
\title{Pattern Matching Operators}
\usage{
x \%like\% pattern

x \%!like\% pattern
}
\arguments{
\item{x}{an object where matches will be sought.}

\item{pattern}{character string containing regular expression to be matched.}
}
\value{
a logical vector or an array of the same dimensions as \code{x}
indicating if each value of \code{x} matched the provided pattern.
}
\description{
Infix operators for pattern matching.
}
\seealso{
\code{grepl}
}
