% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/like_replace.R
\name{like_replace}
\alias{like_replace}
\alias{\%like\%<-}
\alias{\%!like\%<-}
\title{Pattern Matching Replace Operators}
\usage{
\%like\%(x, pattern) <- value

\%!like\%(x, pattern) <- value
}
\arguments{
\item{x}{an object where matches will be sought.}

\item{pattern}{character string containing regular expression to be matched.}

\item{value}{replacement value.}
}
\value{
\code{x} with pattern-matched values replaced with \code{value}.
}
\description{
Infix operators for replacing values matched by a pattern.
}
\seealso{
\code{\%like\%}, code{grepl}
}
