% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_like.R
\name{like_variants}
\alias{like_variants}
\alias{\%like\%}
\alias{\%like\%<-}
\alias{\%!like\%}
\alias{\%!like\%<-}
\title{like variants}
\usage{
x \%like\% pattern

\%like\%(x, pattern) <- value

x \%!like\% pattern

\%!like\%(x, pattern) <- value
}
\arguments{
\item{x}{character or factor vector}

\item{pattern}{pattern to be matched}

\item{value}{replacement value}
}
\description{
like variants
}
