% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_detect.R
\name{out}
\alias{out}
\alias{\%out\%}
\title{Detect values that don't match}
\usage{
x \%out\% table
}
\arguments{
\item{x}{vector of values to be matched.}

\item{table}{vector or list to be matched against.}
}
\value{
a logical vector or of the same length as \code{x}
indicating if each value of \code{x} is within the defined subset.
}
\description{
\code{\%out\%} is the negation of \code{\%in\%}, so \code{x \%out\% y} is equivalent to \code{! x \%in\% y}.
}
\examples{
iris$Species \%in\% c("setosa", "versicolor")
iris$Species \%out\% c("setosa", "versicolor")

# difference in behaviour with dimensions when compared to \%in\%
iris[1:10,] \%in\% "setosa"
iris[1:10,] == "setosa"
iris[1:10,] \%in{}\% "setosa"

# difference in behaviour with missing values when compared to \%in\%
x <- c(1,2,3,NA,4)
x \%in\% c(1,2,3)
x \%in{}\% c(1,2,3)

# other interval oparators
x <- 1:10
x \%in[]\% c(3,7)
x \%in()\% c(3,7)
x \%in(]\% c(3,7)
x \%in[)\% c(3,7)
x \%out[]\% c(3,7)

# when more than 2 numbers are provided for the interval - range is used
x <- 1:10
all.equal(x \%in[]\% c(2,4), x \%in[]\% c(2,3,4))
all.equal(x \%in[]\% c(2,4), x \%in[]\% range(c(2,3,4)))

# matching according to regular expressions
iris$Species \%in~\% "^v"
iris$Species \%in~f\% "^v"
iris$Species \%in~f\% "versicolor"
iris$Species \%in~f\% c("versicolor", "virginica")

}
\seealso{
\code{\%in\%}
}
