% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_interval.R
\name{range_ops}
\alias{range_ops}
\alias{\%in()\%}
\alias{\%in()\%<-}
\alias{\%subset()\%}
\alias{\%!in()\%}
\alias{\%!in()\%<-}
\alias{\%!subset()\%}
\alias{\%in(]\%}
\alias{\%in(]\%<-}
\alias{\%subset(]\%}
\alias{\%!in(]\%}
\alias{\%!in(]\%<-}
\alias{\%!subset(]\%}
\alias{\%in[)\%}
\alias{\%in[)\%<-}
\alias{\%subset[)\%}
\alias{\%!in[)\%}
\alias{\%!in[)\%<-}
\alias{\%!subset[)\%}
\alias{\%in[]\%}
\alias{\%in[]\%<-}
\alias{\%subset[]\%}
\alias{\%!in[]\%}
\alias{\%!in[]\%<-}
\alias{\%!subset[]\%}
\title{Range operators}
\usage{
x \%in()\% range

\%in()\%(x, range) <- value

x \%subset()\% range

x \%!in()\% range

\%!in()\%(x, range) <- value

x \%!subset()\% range

x \%in(]\% range

\%in(]\%(x, range) <- value

x \%subset(]\% range

x \%!in(]\% range

\%!in(]\%(x, range) <- value

x \%!subset(]\% range

x \%in[)\% range

\%in[)\%(x, range) <- value

x \%subset[)\% range

x \%!in[)\% range

\%!in[)\%(x, range) <- value

x \%!subset[)\% range

x \%in[]\% range

\%in[]\%(x, range) <- value

x \%subset[]\% range

x \%!in[]\% range

\%!in[]\%(x, range) <- value

x \%!subset[]\% range
}
\arguments{
\item{x}{an object compatible with comparison operators}

\item{range}{an object compatible with `base::range()`, typically a numeric
vector of size 2 or more, or a list or data frame containing numeric elements.}
}
\description{
Range operators
}
\examples{
vec <- c(0, 3, 6 ,9, 12)
# given a range
vec \%in[]\% c(3, 9)
vec \%in[)\% c(3, 9)
vec \%!in[]\% c(3 ,9)
# range is applied on rhs so we can provide a longer or unsorted vector
vec \%in[]\% c(3,2,7,4)
# range works on data frames or lists as long as they contain only numeric elements
vec \%in[]\% iris[1:4]
# subset
vec \%subset[]\% c(2,7)
# replacement
vec \%in[]\% c(3, 7) <- 8
vec
vec \%in[]\% c(3, 10) <- min
vec
vec \%in[]\% c(-1,5) <- ~.+50
vec
}
