% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_alias.R
\name{set_alias}
\alias{set_alias}
\alias{\%intersect\%}
\alias{\%setdiff\%}
\alias{\%union\%}
\alias{\%setequal\%}
\title{Infix Set Aliases}
\usage{
x \%intersect\% y

x \%setdiff\% y

x \%union\% y

x \%setequal\% y
}
\arguments{
\item{x}{vectors (of the same mode) containing a sequence
    of items (conceptually) with no duplicated values.}

\item{y}{vectors (of the same mode) containing a sequence
    of items (conceptually) with no duplicated values.}
}
\value{
Same result as the corresponding set operation implemented in R.
}
\description{
Alternative infix forms for set operations.
}
\seealso{
\code{intersect}, \code{union}, \code{setdiff}, \code{setequal}
}
