% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_check.R
\name{in_check}
\alias{in_check}
\alias{\%in{}\%}
\alias{\%!in{}\%}
\alias{\%in[]\%}
\alias{\%!in[]\%}
\alias{\%in()\%}
\alias{\%!in()\%}
\alias{\%in(]\%}
\alias{\%!in(]\%}
\alias{\%in[)\%}
\alias{\%!in[)\%}
\alias{\%in~\%}
\alias{\%!in~\%}
\title{Matching Values and Intervals}
\usage{
x \%in{}\% table

x \%!in{}\% table

x \%in[]\% interval

x \%!in[]\% interval

x \%in()\% interval

x \%!in()\% interval

x \%in(]\% interval

x \%!in(]\% interval

x \%in[)\% interval

x \%!in[)\% interval

x \%in~\% pattern

x \%!in~\% pattern
}
\arguments{
\item{x}{vector or array of values to be matched.}

\item{table}{vector or list to be matched against.}

\item{interval}{numeric vector defining a range to be matched against.}

\item{pattern}{character string containing regular expression to be matched.}
}
\value{
a logical vector or an array of the same dimensions as \code{x}
indicating if each value of \code{x} is within the defined subset.
}
\description{
Operators for checking which values are within a given interval or set.
}
\details{
Parentheses of each operator define the type of matching template:\cr
\code{\%in\{\}\%} - set of all elements that will be matched against\cr
\code{\%in[]\%} - closed numeric interval\cr
\code{\%in()\%} - open numeric interval\cr
\code{\%in(]\%} - interval that is is open on the left and closed on the right\cr
\code{\%in[)\%} - interval that is closed on the left and open on the right\cr
\code{\%in~\%} - using a regular expression\cr
Operators starting with a bang (i.e. \code{\%!in\{\}\%}) specify negation.

The \code{\%in\{\}\%} operator is a reimplementation of \code{\%in\%} that preserves the dimensions of \code{x} (see examples).
Operations using numeric intervals are a convenient short hand form of\cr\code{x > interval[1] & x < interval[2]}.

\code{interval} can be specified as a numeric vector of any length.
In such cases \code{range} will be used in order to obtain the endpoints.
}
\seealso{
\code{\%in\%}
}
