% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_subset.R
\name{in_subset}
\alias{in_subset}
\alias{\%[in{}\%}
\alias{\%[out{}\%}
\alias{\%[in[]\%}
\alias{\%[out[]\%}
\alias{\%[in()\%}
\alias{\%[out()\%}
\alias{\%[in(]\%}
\alias{\%[out(]\%}
\alias{\%[in[)\%}
\alias{\%[out[)\%}
\alias{\%[in~\%}
\alias{\%[out~\%}
\title{Subsetting Values and Intervals}
\usage{
x \%in{}\% table <- value

x \%out{}\% table <- value

x \%in[]\% interval <- value

x \%out[]\% interval <- value

x \%in()\% interval <- value

x \%out()\% interval <- value

x \%in(]\% interval <- value

x \%out(]\% interval <- value

x \%in[)\% interval <- value

x \%out[)\% interval <- value

x \%in~\% pattern <- value

x \%out~\% pattern <- value
}
\arguments{
\item{x}{vector or array of values to be matched.}

\item{table}{vector or list to be matched against.}

\item{interval}{numeric vector defining a range to be matched against.
\code{as.numeric()} is used on lists, it will succeed on flat lists but fail on
nested lists}

\item{pattern}{pattern to be matched against.}
}
\value{
\code{x} with specified values replaced with \code{value}.
}
\description{
Operators for replacing values within a given interval or set.
}
\seealso{
\code{\%in{}\%}
}
